refactor: move createPropertyMappingResolver to common place
